+++ packages/odata-service-inquirer/test/unit/index-api.test.ts
@@ -21,7 +21,7 @@ jest.mock('../../src/prompts/datasources/sap-system/system-selection', () => ({
 jest.mock('@sap-ux/store', () => ({
     __esModule: true, // Workaround for spyOn TypeError: Jest cannot redefine property
     ...jest.requireActual('@sap-ux/store'),
-    SystemService: jest.fn().mockImplementation(() => ({
+    getService: jest.fn().mockImplementation(() => ({
         getAll: jest.fn().mockResolvedValue([
             {
                 name: 'storedSystem1',
@@ -58,6 +58,10 @@ jest.mock('@sap-ux/store', () => ({
 describe('API tests', () => {
     beforeEach(() => {
         jest.restoreAllMocks();
+    });
+
+    afterEach(() => {
+        jest.clearAllMocks();
     });
 
     test('getPrompts', async () => {
@@ -120,7 +124,6 @@ describe('API tests', () => {
 
     test('getPrompts, i18n is loaded', async () => {
         const { prompts: questions } = await getPrompts(undefined, undefined, true, undefined, true);
-
         expect(questions).toMatchSnapshot();
     });
 });